Starting to put docker based infrastructure into place
