+++ .gitignore
@@ -11,3 +11,4 @@ venv/
 **/.DS_Store
 .DS_Store/
 .pytest_cache/
+pgdata/
