+++ README.md
@@ -7,4 +7,7 @@
 [Example FastAPI](https://github.com/edkrueger/sars-fastapi)  
 [Another GraphQl, FastAPI Example](https://github.com/cuongld2/graphql_blog_service)  
 [Another GraphQl, FastAPI Example](https://testdriven.io/blog/fastapi-graphql/)  
-[Another GraphQl, FastAPI Example, Source](https://github.com/moluwole/fastapi-graphql)
+[Another GraphQl, FastAPI Example, Source](https://github.com/moluwole/fastapi-graphql)  
+[Terraform and LetsEncrypt](https://dev.to/meseta/automating-fetching-of-wildcard-letsencrypt-https-certificates-for-your-domain-li6)  
+[Docker Best Practices](https://cloud.google.com/architecture/best-practices-for-building-containers)  
+[FastAPI and Docker](https://fastapi.tiangolo.com/deployment/docker/)
