--- alembic/versions/0916d5ee2371_added_watershed_id_to_wgew_raingage_.py
@@ -1,24 +1,0 @@
-"""Added watershed_id to wgew raingage model
-
-Revision ID: 0916d5ee2371
-Revises: 4f7395fdf825
-Create Date: 2021-04-25 13:12:02.948453
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = '0916d5ee2371'
-down_revision = '4f7395fdf825'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    pass
-
-
-def downgrade():
-    pass
