--- alembic/versions/1019c04a7a22_modified_srer_raingage_and_precip_event_.py
@@ -1,28 +1,0 @@
-"""Modified SRER raingage and precip event models
-
-Revision ID: 1019c04a7a22
-Revises: 43513479bb16
-Create Date: 2021-04-30 11:59:28.215546
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = '1019c04a7a22'
-down_revision = '43513479bb16'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
