--- alembic/versions/13280c8beece_added_err_to_wgew_raingage_model.py
@@ -1,28 +1,0 @@
-"""Added err to wgew raingage model
-
-Revision ID: 13280c8beece
-Revises: 297109d465c7
-Create Date: 2021-04-25 13:19:16.809867
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = '13280c8beece'
-down_revision = '297109d465c7'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('wgew_raingage', sa.Column('err', sa.Numeric(precision=5, scale=1), nullable=True))
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('wgew_raingage', 'err')
-    # ### end Alembic commands ###
