--- alembic/versions/233dbe245b58_added_time_est_to_wgew_precip_event_.py
@@ -1,28 +1,0 @@
-"""Added time_est to wgew precip event model
-
-Revision ID: 233dbe245b58
-Revises: d82d9243fb26
-Create Date: 2021-04-25 14:02:09.650596
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = '233dbe245b58'
-down_revision = 'd82d9243fb26'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('wgew_precipevent', sa.Column('time_est', sa.String(length=2), nullable=True))
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('wgew_precipevent', 'time_est')
-    # ### end Alembic commands ###
