--- alembic/versions/297109d465c7_added_elevation_to_wgew_raingage_model.py
@@ -1,28 +1,0 @@
-"""Added elevation to wgew raingage model
-
-Revision ID: 297109d465c7
-Revises: ed0355618895
-Create Date: 2021-04-25 13:18:17.783988
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = '297109d465c7'
-down_revision = 'ed0355618895'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('wgew_raingage', sa.Column('elevation', sa.Integer(), nullable=True))
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('wgew_raingage', 'elevation')
-    # ### end Alembic commands ###
