--- alembic/versions/34fe8c5bea4c_added_east_and_north_to_wgew_raingage_.py
@@ -1,30 +1,0 @@
-"""Added east and north to wgew raingage model
-
-Revision ID: 34fe8c5bea4c
-Revises: 13280c8beece
-Create Date: 2021-04-25 13:21:33.222459
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = '34fe8c5bea4c'
-down_revision = '13280c8beece'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('wgew_raingage', sa.Column('east', sa.Integer(), nullable=True))
-    op.add_column('wgew_raingage', sa.Column('north', sa.Integer(), nullable=True))
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('wgew_raingage', 'north')
-    op.drop_column('wgew_raingage', 'east')
-    # ### end Alembic commands ###
