--- alembic/versions/43513479bb16_added_srer_raingage_and_precip_event_.py
@@ -1,40 +1,0 @@
-"""Added SRER raingage and precip event models
-
-Revision ID: 43513479bb16
-Revises: 027d5ec052c2
-Create Date: 2021-04-30 11:27:00.570871
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = '43513479bb16'
-down_revision = '027d5ec052c2'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('srer_precipevent',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('year', sa.Integer(), nullable=True),
-    sa.Column('month', sa.Integer(), nullable=True),
-    sa.Column('precip', sa.Integer(), nullable=True),
-    sa.Column('created', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.Column('updated', sa.DateTime(timezone=True), nullable=True),
-    sa.Column('raingage_id', sa.Integer(), nullable=True),
-    sa.ForeignKeyConstraint(['raingage_id'], ['srer_raingage.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_srer_precipevent_id'), 'srer_precipevent', ['id'], unique=False)
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index(op.f('ix_srer_precipevent_id'), table_name='srer_precipevent')
-    op.drop_table('srer_precipevent')
-    # ### end Alembic commands ###
