--- alembic/versions/46a22901d2fd_wgew_one_to_many_on_raingages_to_precip_.py
@@ -1,30 +1,0 @@
-"""WGEW one-to-many on raingages to precip events
-
-Revision ID: 46a22901d2fd
-Revises: b39cc74d096c
-Create Date: 2021-04-24 16:25:13.398895
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = '46a22901d2fd'
-down_revision = 'b39cc74d096c'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('wgew_precipevent', sa.Column('raingage_id', sa.Integer(), nullable=True))
-    op.create_foreign_key(None, 'wgew_precipevent', 'wgew_raingage', ['raingage_id'], ['id'])
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_constraint(None, 'wgew_precipevent', type_='foreignkey')
-    op.drop_column('wgew_precipevent', 'raingage_id')
-    # ### end Alembic commands ###
