--- alembic/versions/4f7395fdf825_added_watershed_id_to_wgew_raingage_.py
@@ -1,24 +1,0 @@
-"""Added watershed_id to wgew raingage model
-
-Revision ID: 4f7395fdf825
-Revises: e339d201931a
-Create Date: 2021-04-25 13:10:59.041212
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = '4f7395fdf825'
-down_revision = 'e339d201931a'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    pass
-
-
-def downgrade():
-    pass
