--- alembic/versions/530c43205f9a_added_latitude_and_longitude_to_wgew_.py
@@ -1,28 +1,0 @@
-"""Added latitude and longitude to wgew raingage model
-
-Revision ID: 530c43205f9a
-Revises: fa38d6ffc817
-Create Date: 2021-04-25 13:16:59.049387
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = '530c43205f9a'
-down_revision = 'fa38d6ffc817'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
