--- alembic/versions/8060b0cf1863_initialization.py
@@ -1,24 +1,0 @@
-"""Initialization
-
-Revision ID: 8060b0cf1863
-Revises: 
-Create Date: 2021-04-24 16:01:56.767227
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = '8060b0cf1863'
-down_revision = None
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    pass
-
-
-def downgrade():
-    pass
