--- alembic/versions/8c6f6bab7fd7_wgew_one_to_many_on_raingages_to_precip_.py
@@ -1,28 +1,0 @@
-"""WGEW one-to-many on raingages to precip events. Update
-
-Revision ID: 8c6f6bab7fd7
-Revises: 46a22901d2fd
-Create Date: 2021-04-24 16:27:28.959637
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = '8c6f6bab7fd7'
-down_revision = '46a22901d2fd'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
