--- alembic/versions/9d2fc1fb141e_added_watershed_id_to_wgew_raingage_.py
@@ -1,28 +1,0 @@
-"""Added watershed_id to wgew raingage model
-
-Revision ID: 9d2fc1fb141e
-Revises: 0916d5ee2371
-Create Date: 2021-04-25 13:12:29.267483
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = '9d2fc1fb141e'
-down_revision = '0916d5ee2371'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('wgew_raingage', sa.Column('watershed_id', sa.String(length=5), nullable=True))
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('wgew_raingage', 'watershed_id')
-    # ### end Alembic commands ###
