--- alembic/versions/a85526ff71f7_modified_srer_raingage_model.py
@@ -1,28 +1,0 @@
-"""Modified SRER raingage model
-
-Revision ID: a85526ff71f7
-Revises: ebb2d7645e00
-Create Date: 2021-04-29 00:54:56.154924
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = 'a85526ff71f7'
-down_revision = 'ebb2d7645e00'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
