--- alembic/versions/a8834437c8f1_adjusted_gage_id_to_int_in_wgew_.py
@@ -1,28 +1,0 @@
-"""Adjusted gage_id to int in wgew raingage model
-
-Revision ID: a8834437c8f1
-Revises: ba9e6a73a17d
-Create Date: 2021-04-25 14:10:29.069164
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = 'a8834437c8f1'
-down_revision = 'ba9e6a73a17d'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
