--- alembic/versions/b39cc74d096c_initialization_to_all_models.py
@@ -1,44 +1,0 @@
-"""Initialization to all models
-
-Revision ID: b39cc74d096c
-Revises: 8060b0cf1863
-Create Date: 2021-04-24 16:05:10.716906
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = 'b39cc74d096c'
-down_revision = '8060b0cf1863'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('wgew_precipevent',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('created', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.Column('updated', sa.DateTime(timezone=True), nullable=True),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_wgew_precipevent_id'), 'wgew_precipevent', ['id'], unique=False)
-    op.create_table('wgew_raingage',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('created', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.Column('updated', sa.DateTime(timezone=True), nullable=True),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_wgew_raingage_id'), 'wgew_raingage', ['id'], unique=False)
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index(op.f('ix_wgew_raingage_id'), table_name='wgew_raingage')
-    op.drop_table('wgew_raingage')
-    op.drop_index(op.f('ix_wgew_precipevent_id'), table_name='wgew_precipevent')
-    op.drop_table('wgew_precipevent')
-    # ### end Alembic commands ###
