--- alembic/versions/ba9e6a73a17d_adjusted_gage_id_to_int_in_wgew_.py
@@ -1,28 +1,0 @@
-"""Adjusted gage_id to int in wgew raingage model
-
-Revision ID: ba9e6a73a17d
-Revises: 233dbe245b58
-Create Date: 2021-04-25 14:09:54.564948
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = 'ba9e6a73a17d'
-down_revision = '233dbe245b58'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
