--- alembic/versions/d82d9243fb26_added_depth_to_wgew_precip_event_model.py
@@ -1,28 +1,0 @@
-"""Added depth to wgew precip event model
-
-Revision ID: d82d9243fb26
-Revises: e9524f56fc8b
-Create Date: 2021-04-25 14:00:47.724509
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = 'd82d9243fb26'
-down_revision = 'e9524f56fc8b'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('wgew_precipevent', sa.Column('depth', sa.Numeric(precision=15, scale=5), nullable=True))
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('wgew_precipevent', 'depth')
-    # ### end Alembic commands ###
