--- alembic/versions/e339d201931a_wgew_one_to_many_on_raingages_to_precip_.py
@@ -1,28 +1,0 @@
-"""WGEW one-to-many on raingages to precip events. Update again
-
-Revision ID: e339d201931a
-Revises: 8c6f6bab7fd7
-Create Date: 2021-04-24 16:38:10.403789
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = 'e339d201931a'
-down_revision = '8c6f6bab7fd7'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
