--- alembic/versions/ed0355618895_added_latitude_and_longitude_to_wgew_.py
@@ -1,30 +1,0 @@
-"""Added latitude and longitude to wgew raingage model
-
-Revision ID: ed0355618895
-Revises: 530c43205f9a
-Create Date: 2021-04-25 13:17:20.961448
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = 'ed0355618895'
-down_revision = '530c43205f9a'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('wgew_raingage', sa.Column('latitude', sa.Numeric(precision=15, scale=5), nullable=True))
-    op.add_column('wgew_raingage', sa.Column('longitude', sa.Numeric(precision=15, scale=5), nullable=True))
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('wgew_raingage', 'longitude')
-    op.drop_column('wgew_raingage', 'latitude')
-    # ### end Alembic commands ###
