--- alembic/versions/f633e16dd3b1_adjusted_watershed_id_to_int_in_wgew_.py
@@ -1,28 +1,0 @@
-"""Adjusted watershed_id to int in wgew raingage model
-
-Revision ID: f633e16dd3b1
-Revises: df72420ae702
-Create Date: 2021-04-25 14:15:32.465213
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = 'f633e16dd3b1'
-down_revision = 'df72420ae702'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
