--- alembic/versions/fa38d6ffc817_added_gage_id_to_wgew_raingage_model.py
@@ -1,28 +1,0 @@
-"""Added gage_id to wgew raingage model
-
-Revision ID: fa38d6ffc817
-Revises: 9d2fc1fb141e
-Create Date: 2021-04-25 13:14:04.606497
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = 'fa38d6ffc817'
-down_revision = '9d2fc1fb141e'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('wgew_raingage', sa.Column('gage_id', sa.String(length=125), nullable=True))
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('wgew_raingage', 'gage_id')
-    # ### end Alembic commands ###
