+++ docker-compose.yml
@@ -1,13 +1,40 @@
 version: '3.9'
 
 services:
-    db:
-        image: postgres:alpine
+    postgres:
         env_file: .env
         ports:
             - 5432:5432
-    adminer:
-        image: adminer
-        # restart: always
+        container_name: eapidb
+        image: postgres:13.3-alpine        
+        volumes:
+            - dbdata:/var/lib/postgresql/data
+        networks:
+            - eapinet
+    app:
+        env_file: .env
         ports:
-            - 8080:8080
+            - 8080:80
+        container_name: eapiapp
+        build:
+            dockerfile: Dockerfile
+            context: .
+        volumes:
+            - $PWD:/eapi
+        networks:
+            - eapinet
+
+networks:
+    eapinet:
+volumes:
+    dbdata:
+
+    # nginx:
+    #     env_file: .env
+    #     container_name: nginx
+    #     build:
+    #         dockerfile: Dockerfile
+    #         context: ./nginx
+    #     restart: "no"
+    #     ports:
+    #         - 8000:80
